+++ .babelrc
@@ -9,5 +9,6 @@
       }
     ],
     "react"
-  ]
+  ],
+  "plugins": ["babel-plugin-transform-object-rest-spread"]
 }
