+++ package.json
@@ -5,10 +5,11 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "lint": "eslint -c .eslintrc.js ./src/**/*.js ./src/**/*.jsx",
+    "lint": "eslint -c .eslintrc.js ./src/**/*.js ./src/**/*.jsx ./server/**/*.js",
     "client-build": "webpack --mode production",
     "client-build-dev": "webpack --mode development",
     "client-start": "webpack --mode development && webpack-dev-server --mode development --open",
+    "create-json": "node server/script.js",
     "server-clean": "rm -rf dist-server && mkdir dist-server",
     "server-build": "npm run server-clean && babel ./server -d ./dist-server",
     "server-start": "npm run server-build && nodemon ./dist-server/app.js",
@@ -33,14 +34,17 @@
     "babel-core": "^6.26.3",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^7.1.5",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
+    "body-parser": "^1.18.3",
     "copy-webpack-plugin": "^4.6.0",
     "css-loader": "^2.1.0",
     "eslint": "^5.14.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-react": "^7.12.4",
+    "file-loader": "^3.0.1",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.1.0",
     "nodemon": "^1.18.10",
@@ -50,6 +54,7 @@
     "webpack-dev-server": "^3.1.14"
   },
   "dependencies": {
+    "express": "^4.16.4",
     "firebase": "^5.8.4",
     "firebase-tools": "^6.4.0",
     "prop-types": "^15.7.2",
